move CPlusPlus to own dir
