+++ packages/quicktype-core/src/language/CPlusPlus/CPlusPlusRenderer.ts
@@ -8,473 +8,47 @@ import {
     withDefault
 } from "collection-utils";
 
-import { anyTypeIssueAnnotation, nullTypeIssueAnnotation } from "../Annotation";
-import { getAccessorName } from "../attributes/AccessorNames";
+import { anyTypeIssueAnnotation, nullTypeIssueAnnotation } from "../../Annotation";
+import { getAccessorName } from "../../attributes/AccessorNames";
+import { enumCaseValues } from "../../attributes/EnumValues";
+import { ConvenienceRenderer, type ForbiddenWordsInfo } from "../../ConvenienceRenderer";
+import { type Declaration } from "../../DeclarationIR";
+import { DependencyName, type Name, type NameStyle, type Namer, funPrefixNamer } from "../../Naming";
+import { type RenderContext } from "../../Renderer";
+import { type OptionValues } from "../../RendererOptions";
+import { type Sourcelike, maybeAnnotated } from "../../Source";
+import { type NamingStyle, makeNameStyle, stringEscape } from "../../support/Strings";
+import { assert, assertNever, defined, numberEnumValues, panic } from "../../support/Support";
+import { type TargetLanguage } from "../../TargetLanguage";
+import { ArrayType, type ClassProperty, ClassType, EnumType, MapType, type Type, UnionType } from "../../Type";
 import {
-    type MinMaxConstraint,
-    minMaxLengthForType,
-    minMaxValueForType,
-    patternForType
-} from "../attributes/Constraints";
-import { enumCaseValues } from "../attributes/EnumValues";
-import { ConvenienceRenderer, type ForbiddenWordsInfo } from "../ConvenienceRenderer";
-import { type Declaration } from "../DeclarationIR";
-import { DependencyName, type Name, type NameStyle, type Namer, funPrefixNamer } from "../Naming";
-import { type RenderContext } from "../Renderer";
+    directlyReachableTypes,
+    isNamedType,
+    matchType,
+    nullableFromUnion,
+    removeNullFromUnion
+} from "../../TypeUtils";
+
+import { keywords } from "./constants";
+import { type cPlusPlusOptions } from "./language";
 import {
-    BooleanOption,
-    EnumOption,
-    type Option,
-    type OptionValues,
-    StringOption,
-    getOptionValues
-} from "../RendererOptions";
-import { type Sourcelike, maybeAnnotated } from "../Source";
-import {
-    type NamingStyle,
-    isAscii,
-    isLetterOrUnderscoreOrDigit,
-    legalizeCharacters,
-    makeNameStyle,
-    stringEscape
-} from "../support/Strings";
-import { assert, assertNever, defined, numberEnumValues, panic } from "../support/Support";
-import { TargetLanguage } from "../TargetLanguage";
-import {
-    ArrayType,
-    type ClassProperty,
-    ClassType,
-    EnumType,
-    MapType,
-    type Type,
-    type TypeKind,
-    UnionType
-} from "../Type";
-import { type FixMeOptionsAnyType, type FixMeOptionsType } from "../types";
-import { directlyReachableTypes, isNamedType, matchType, nullableFromUnion, removeNullFromUnion } from "../TypeUtils";
-
-const pascalValue: [string, NamingStyle] = ["pascal-case", "pascal"];
-const underscoreValue: [string, NamingStyle] = ["underscore-case", "underscore"];
-const camelValue: [string, NamingStyle] = ["camel-case", "camel"];
-const upperUnderscoreValue: [string, NamingStyle] = ["upper-underscore-case", "upper-underscore"];
-const pascalUpperAcronymsValue: [string, NamingStyle] = ["pascal-case-upper-acronyms", "pascal-upper-acronyms"];
-const camelUpperAcronymsValue: [string, NamingStyle] = ["camel-case-upper-acronyms", "camel-upper-acronyms"];
-
-export const cPlusPlusOptions = {
-    typeSourceStyle: new EnumOption(
-        "source-style",
-        "Source code generation type,  whether to generate single or multiple source files",
-        [
-            ["single-source", true],
-            ["multi-source", false]
-        ],
-        "single-source",
-        "secondary"
-    ),
-    includeLocation: new EnumOption(
-        "include-location",
-        "Whether json.hpp is to be located globally or locally",
-        [
-            ["local-include", true],
-            ["global-include", false]
-        ],
-        "local-include",
-        "secondary"
-    ),
-    codeFormat: new EnumOption(
-        "code-format",
-        "Generate classes with getters/setters, instead of structs",
-        [
-            ["with-struct", false],
-            ["with-getter-setter", true]
-        ],
-        "with-getter-setter"
-    ),
-    wstring: new EnumOption(
-        "wstring",
-        "Store strings using Utf-16 std::wstring, rather than Utf-8 std::string",
-        [
-            ["use-string", false],
-            ["use-wstring", true]
-        ],
-        "use-string"
-    ),
-    westConst: new EnumOption(
-        "const-style",
-        "Put const to the left/west (const T) or right/east (T const)",
-        [
-            ["west-const", true],
-            ["east-const", false]
-        ],
-        "west-const"
-    ),
-    justTypes: new BooleanOption("just-types", "Plain types only", false),
-    namespace: new StringOption("namespace", "Name of the generated namespace(s)", "NAME", "quicktype"),
-    enumType: new StringOption("enum-type", "Type of enum class", "NAME", "int", "secondary"),
-    typeNamingStyle: new EnumOption<NamingStyle>("type-style", "Naming style for types", [
-        pascalValue,
-        underscoreValue,
-        camelValue,
-        upperUnderscoreValue,
-        pascalUpperAcronymsValue,
-        camelUpperAcronymsValue
-    ]),
-    memberNamingStyle: new EnumOption<NamingStyle>("member-style", "Naming style for members", [
-        underscoreValue,
-        pascalValue,
-        camelValue,
-        upperUnderscoreValue,
-        pascalUpperAcronymsValue,
-        camelUpperAcronymsValue
-    ]),
-    enumeratorNamingStyle: new EnumOption<NamingStyle>("enumerator-style", "Naming style for enumerators", [
-        upperUnderscoreValue,
-        underscoreValue,
-        pascalValue,
-        camelValue,
-        pascalUpperAcronymsValue,
-        camelUpperAcronymsValue
-    ]),
-    boost: new BooleanOption("boost", "Require a dependency on boost. Without boost, C++17 is required", true),
-    hideNullOptional: new BooleanOption("hide-null-optional", "Hide null value for optional field", false)
-};
-
-export class CPlusPlusTargetLanguage extends TargetLanguage {
-    public constructor(displayName = "C++", names: string[] = ["c++", "cpp", "cplusplus"], extension = "cpp") {
-        super(displayName, names, extension);
-    }
-
-    protected getOptions(): Array<Option<FixMeOptionsAnyType>> {
-        return [
-            cPlusPlusOptions.justTypes,
-            cPlusPlusOptions.namespace,
-            cPlusPlusOptions.codeFormat,
-            cPlusPlusOptions.wstring,
-            cPlusPlusOptions.westConst,
-            cPlusPlusOptions.typeSourceStyle,
-            cPlusPlusOptions.includeLocation,
-            cPlusPlusOptions.typeNamingStyle,
-            cPlusPlusOptions.memberNamingStyle,
-            cPlusPlusOptions.enumeratorNamingStyle,
-            cPlusPlusOptions.enumType,
-            cPlusPlusOptions.boost,
-            cPlusPlusOptions.hideNullOptional
-        ];
-    }
-
-    public get supportsUnionsWithBothNumberTypes(): boolean {
-        return true;
-    }
-
-    public get supportsOptionalClassProperties(): boolean {
-        return true;
-    }
-
-    protected makeRenderer(renderContext: RenderContext, untypedOptionValues: FixMeOptionsType): CPlusPlusRenderer {
-        return new CPlusPlusRenderer(this, renderContext, getOptionValues(cPlusPlusOptions, untypedOptionValues));
-    }
-}
-
-function constraintsForType(t: Type):
-    | {
-          minMax?: MinMaxConstraint;
-          minMaxLength?: MinMaxConstraint;
-          pattern?: string;
-      }
-    | undefined {
-    const minMax = minMaxValueForType(t);
-    const minMaxLength = minMaxLengthForType(t);
-    const pattern = patternForType(t);
-    if (minMax === undefined && minMaxLength === undefined && pattern === undefined) return undefined;
-    return { minMax, minMaxLength, pattern };
-}
-
-const legalizeName = legalizeCharacters(cp => isAscii(cp) && isLetterOrUnderscoreOrDigit(cp));
-
-const keywords = [
-    "alignas",
-    "alignof",
-    "and",
-    "and_eq",
-    "asm",
-    "atomic_cancel",
-    "atomic_commit",
-    "atomic_noexcept",
-    "auto",
-    "bitand",
-    "bitor",
-    "bool",
-    "break",
-    "case",
-    "catch",
-    "char",
-    "char16_t",
-    "char32_t",
-    "class",
-    "compl",
-    "concept",
-    "const",
-    "constexpr",
-    "const_cast",
-    "continue",
-    "co_await",
-    "co_return",
-    "co_yield",
-    "decltype",
-    "default",
-    "delete",
-    "do",
-    "double",
-    "dynamic_cast",
-    "else",
-    "enum",
-    "explicit",
-    "export",
-    "extern",
-    "false",
-    "float",
-    "for",
-    "friend",
-    "goto",
-    "if",
-    "import",
-    "inline",
-    "int",
-    "long",
-    "module",
-    "mutable",
-    "namespace",
-    "new",
-    "noexcept",
-    "not",
-    "not_eq",
-    "nullptr",
-    "operator",
-    "or",
-    "or_eq",
-    "private",
-    "protected",
-    "public",
-    "register",
-    "reinterpret_cast",
-    "requires",
-    "return",
-    "short",
-    "signed",
-    "sizeof",
-    "static",
-    "static_assert",
-    "static_cast",
-    "struct",
-    "switch",
-    "synchronized",
-    "template",
-    "this",
-    "thread_local",
-    "throw",
-    "true",
-    "try",
-    "typedef",
-    "typeid",
-    "typename",
-    "union",
-    "unsigned",
-    "using",
-    "virtual",
-    "void",
-    "volatile",
-    "wchar_t",
-    "while",
-    "xor",
-    "xor_eq",
-    "override",
-    "final",
-    "transaction_safe",
-    "transaction_safe_dynamic",
-    "NULL"
-];
-
-/// Type to use as an optional if cycle breaking is required
-const optionalAsSharedType = "std::shared_ptr";
-/// Factory to use when creating an optional if cycle breaking is required
-const optionalFactoryAsSharedType = "std::make_shared";
-
-/**
- * To be able to support circles in multiple files -
- * e.g. class#A using class#B using class#A (obviously not directly,
- * but in vector or in variant) we can forward declare them;
- */
-export enum IncludeKind {
-    ForwardDeclare = "ForwardDeclare",
-    Include = "Include"
-}
-
-// FIXME: make these string enums eventually
-export enum GlobalNames {
-    ClassMemberConstraints = 1,
-    ClassMemberConstraintException = 2,
-    ValueTooLowException = 3,
-    ValueTooHighException = 4,
-    ValueTooShortException = 5,
-    ValueTooLongException = 6,
-    InvalidPatternException = 7,
-    CheckConstraint = 8
-}
-
-// FIXME: make these string enums eventually
-export enum MemberNames {
-    MinIntValue = 1,
-    GetMinIntValue = 2,
-    SetMinIntValue = 3,
-    MaxIntValue = 4,
-    GetMaxIntValue = 5,
-    SetMaxIntValue = 6,
-    MinDoubleValue = 7,
-    GetMinDoubleValue = 8,
-    SetMinDoubleValue = 9,
-    MaxDoubleValue = 10,
-    GetMaxDoubleValue = 11,
-    SetMaxDoubleValue = 12,
-    MinLength = 13,
-    GetMinLength = 14,
-    SetMinLength = 15,
-    MaxLength = 16,
-    GetMaxLength = 17,
-    SetMaxLength = 18,
-    Pattern = 19,
-    GetPattern = 20,
-    SetPattern = 21
-}
-
-interface ConstraintMember {
-    cppConstType?: string;
-    cppType: string;
-    getter: MemberNames;
-    name: MemberNames;
-    setter: MemberNames;
-}
-
-export interface IncludeRecord {
-    kind: IncludeKind | undefined /** How to include that */;
-    typeKind: TypeKind | undefined /** What exactly to include */;
-}
-
-export interface TypeRecord {
-    forceInclude: boolean;
-    level: number;
-    name: Name;
-    type: Type;
-    variant: boolean;
-}
-
-/**
- * We map each and every unique type to a include kind, e.g. how
- * to include the given type
- */
-export type IncludeMap = Map<string, IncludeRecord>;
-
-export interface TypeContext {
-    inJsonNamespace: boolean;
-    needsForwardIndirection: boolean;
-    needsOptionalIndirection: boolean;
-}
-
-interface StringType {
-    createStringLiteral: (inner: Sourcelike) => Sourcelike;
-    emitHelperFunctions: () => void;
-    getConstType: () => string;
-    getRegex: () => string;
-    getSMatch: () => string;
-    getType: () => string;
-    wrapEncodingChange: (
-        qualifier: Sourcelike[],
-        fromType: Sourcelike,
-        toType: Sourcelike,
-        inner: Sourcelike
-    ) => Sourcelike;
-    wrapToString: (inner: Sourcelike) => Sourcelike;
-}
-
-function addQualifier(qualifier: Sourcelike, qualified: Sourcelike[]): Sourcelike[] {
-    if (qualified.length === 0) {
-        return [];
-    }
-
-    return [qualifier, qualified];
-}
-
-class WrappingCode {
-    public constructor(
-        private readonly start: Sourcelike[],
-        private readonly end: Sourcelike[]
-    ) {}
-
-    public wrap(qualifier: Sourcelike, inner: Sourcelike): Sourcelike {
-        return [addQualifier(qualifier, this.start), inner, this.end];
-    }
-}
-
-class BaseString {
-    public _stringType: string;
-
-    public _constStringType: string;
-
-    public _smatch: string;
-
-    public _regex: string;
-
-    public _stringLiteralPrefix: string;
-
-    public _toString: WrappingCode;
-
-    public _encodingClass: Sourcelike;
-
-    public _encodingFunction: Sourcelike;
-
-    public constructor(
-        stringType: string,
-        constStringType: string,
-        smatch: string,
-        regex: string,
-        stringLiteralPrefix: string,
-        toString: WrappingCode,
-        encodingClass: string,
-        encodingFunction: string
-    ) {
-        this._stringType = stringType;
-        this._constStringType = constStringType;
-        this._smatch = smatch;
-        this._regex = regex;
-        this._stringLiteralPrefix = stringLiteralPrefix;
-        this._toString = toString;
-        this._encodingClass = encodingClass;
-        this._encodingFunction = encodingFunction;
-    }
-
-    public getType(): string {
-        return this._stringType;
-    }
-
-    public getConstType(): string {
-        return this._constStringType;
-    }
-
-    public getSMatch(): string {
-        return this._smatch;
-    }
-
-    public getRegex(): string {
-        return this._regex;
-    }
-
-    public createStringLiteral(inner: Sourcelike): Sourcelike {
-        return [this._stringLiteralPrefix, '"', inner, '"'];
-    }
-
-    public wrapToString(inner: Sourcelike): Sourcelike {
-        return this._toString.wrap([], inner);
-    }
-}
+    BaseString,
+    type ConstraintMember,
+    GlobalNames,
+    IncludeKind,
+    type IncludeMap,
+    type IncludeRecord,
+    MemberNames,
+    type StringType,
+    type TypeContext,
+    type TypeRecord,
+    WrappingCode,
+    addQualifier,
+    constraintsForType,
+    legalizeName,
+    optionalAsSharedType,
+    optionalFactoryAsSharedType
+} from "./utils";
 
 export class CPlusPlusRenderer extends ConvenienceRenderer {
     /**
