+++ packages/quicktype-core/src/language/CPlusPlus/index.ts
@@ -1,0 +1,2 @@
+export { CPlusPlusTargetLanguage, cPlusPlusOptions } from "./language";
+export { CPlusPlusRenderer } from "./CPlusPlusRenderer";
